T11683 RandomWiki: Stop returning closed wikis, add inactive option
For Special:RandomWiki we never want to return a closed wiki
or a private one, so just remove the code to do so. (If someone
later wants to do randomly select closed wikis for some reason
we can just give them a copy of the cw_wikis table and some
polyhedral dice.)

Also all of the code was there for looking at inactive wikis
or active, so I enabled that code because this might be interesting.
People might only want to find active projects or try to save inactive ones.