T11683 RandomWiki: Let the DB pick a random wiki, use wiki_url field if available
To reduce traffic, let's try doing one query with a random response
instead of two queries with lots of information discarded by the
application.  It may be slower (I doubt it) but it will definitely
reduce memory and data transfer.  (Alternately we could try caching
cw_wikis and choosing random on the application side, but the list
of wikis will only get bigger.

Apparently wiki_url is still nullable for now, but that should be
the field we always use to redirect.